#11: Configure the metadata prefix per repository
Previously the metadata prefix was an application-wide setting, which resulted in the need for separate configurations for each repository that had a different metadata prefix. The metadata prefix has been moved into a per repository setting, and the harvest will fail if the metadata prefix is not configured.

The harvester has also been configured to be command line application, rather than running as a web server.