#409: Export repository configuration to pipeline.json, refactor data model to include configuration needed for downstream components.
This commit adds support for exporting metadata about the current harvest to a JSON file called pipeline.json. This file describes metadata about the repository that is needed by downstream components such as the validator and the indexer.

This commit also simplifies the configuration model of the harvester by introducing a MetadataFormat record that combines the setSpec and metadata prefix, as well as adding information about the DDI version and the validation profile to be used.

See https://docs.google.com/document/d/110rvywK-Y_NbmzgT_k3lszXjo9D831ta for the design document.