Define a custom style for mdl
By default the rule for line length (MD013) enforces to use less than
80 characters for tables. This typically leads to either non-readable
tables or non-rendable markdown.

Few excerpts from "Markdown philosophy" at
https://daringfireball.net/projects/markdown/syntax#philosophy

"Markdown is intended to be as easy-to-read and easy-to-write as is
feasible."
[...]
"Readability, however, is emphasized above all else. A
Markdown-formatted document should be publishable as-is, as plain
text, without looking like it’s been marked up with tags or formatting
instructions."

The Github markdown renderer is totally fine with tables having a line
length beyond 80 characters. However the SQAaaS Bronze badge requires
to comply to all rules of the markdownlint. Therefore this commit
defines a custom style for markdownlint (mdl) that passes tables=false
to the rule 'MD013'. The style does no other customizations to the
default style.

Add file .mdlrc to configure mdl to use the customized style.

Add file .mdl_custom_style.rb which defines the customized style, by
passing tables=false to rule 'MD013'.