#736: Introduce a new configuration format for repositories
One of the problems with the current format is that the method for configuring a repository with multiple metadata prefixes and setSpec cannot have different codes or names.

The new configuration format does not have nested arrays. Instead, the user should create multiple repository entries in order for multiple metadata prefixes and sets for each repository.

This commit implements the new pipeline definitions defined in https://github.com/cessda/cessda.cdc.versions/pull/736.